# perf(NODE-6344): Improve ObjectId.isValid(string) performance