### Description

This MR improves performance of using `ObjectId.isValid('some string')` to check if a string is a valid hex representation of an ObjectId. We frequently call this function to check if a user's input is valid (for example, when creating a document that links to other document(s)).

### Performance tests

| Test                        | Ops/s (Main)         | Ops/s (New)          | % Improvement   |
|-----------------------------|----------------------|----------------------|-----------------|
| bson#new ObjectId(string)    | 3,173,643 ops/sec    | 3,299,153 ops/sec    | 3.95%           |
| bson#isValid                | 5,513,041 ops/sec    | 22,529,639 ops/sec   | 308.66%         |
| bson#isValid uppercase      | 5,498,098 ops/sec    | 21,776,208 ops/sec   | 296.07%         |
| bson#isValid invalid end       | 131,657 ops/sec      | 22,592,719 ops/sec   | 17,060.24%      |
| bson#isValid invalid start  | 132,751 ops/sec      | 152,311,298 ops/sec  | 114,634.47%     |

There is a drastic performance improvement when checking for invalid strings since throwing/catching errors is expensive.

#### What is changing?

##### Is there new documentation needed for these changes?

None

#### What is the motivation for this change?

<!-- If this is a bug, it helps to describe the current behavior and a clear outline of the expected behavior -->
<!-- If this is a feature, it helps to describe the new use case enabled by this change -->

<!--
Contributors!
First of all, thank you so much!!
If you haven't already, it would greatly help the team review this work in a timely manner if you create a JIRA ticket to track this PR.
You can do that here: https://jira.mongodb.org/projects/NODE
-->

### Release Highlight

<!-- RELEASE_HIGHLIGHT_START -->

### Fill in title or leave empty for no highlight

<!-- RELEASE_HIGHLIGHT_END -->

### Double check the following

- [x] Ran `npm run check:lint` script
- [x] Self-review completed using the [steps outlined here](https://github.com/mongodb/node-mongodb-native/blob/HEAD/CONTRIBUTING.md#reviewer-guidelines)
- [x] PR title follows the [correct format](https://www.conventionalcommits.org/en/v1.0.0/): `type(NODE-xxxx)[!]: description`
  - Example: `feat(NODE-1234)!: rewriting everything in coffeescript`
- [x] Changes are covered by tests
- [x] New TODOs have a related JIRA ticket