@baileympearson 2024-08-27
Hey @SeanReece - I don't believe we have benchmarks for this method.  Could you share your benchmarking script so I can test and reproduce?