@SeanReece 2024-08-27
@baileympearson Sure thing. I just threw together a simple benchmark locally since I can't get bson-bench to run locally. If you pull this branch locally, change the name in `package.json` to `bson-isvalid` and run an `npm run build` then you can import bson main and this branch alongside each other.

```
var Benchmark = require("benchmark");
const bson = require("bson");
const bsonVAL = require("bson-isvalid");

var suite = new Benchmark.Suite();

suite
  // Test calling isValid with a valid objectId hex
  .add("bson#isValid", function () {
    bson.ObjectId.isValid("6683fe22cd402749418c7e2b");
  })
  .add("bsonVAL#isValid", function () {
    bsonVAL.ObjectId.isValid("6683fe22cd402749418c7e2b");
  })

   // Test calling isValid with an INVALID objectId hex
  .add("bson#isValid with invalid string", function () {
    bson.ObjectId.isValid("6683fE22cd402749418c7e2x");
  })
  .add("bsonVAL#isValid with invalid string", function () {
    bsonVAL.ObjectId.isValid("6683fE22cd402749418c7e2x");
  })
  .on("cycle", function (event) {
    console.log(String(event.target));
  })
  .on("complete", function () {
    console.log("Fastest is " + this.filter("fastest").map("name"));
  })
  .run({ async: true });
```