@baileympearson 2024-08-27
@SeanReece Thanks - I see the same behavior.

I'll bring this up with the team later this week.