@durran 2024-09-05
@SeanReece just as the other PR for isValid I've rebased this one as well.